+++ src/mechanisms/scram/server.rs
@@ -4,7 +4,7 @@ use crate::mechanisms::scram::parser::{
     ClientFinal, ClientFirstMessage, GS2CBindFlag, ServerErrorValue, ServerFinal, ServerFirst,
 };
 use crate::mechanisms::scram::tools::{compute_signatures, generate_nonce, DOutput};
-use crate::session::{MechanismData, State};
+use crate::session::{MechanismData, MessageSent, State};
 use crate::vectored_io::VectoredWriter;
 use digest::crypto_common::BlockSizeUser;
 use digest::generic_array::GenericArray;
@@ -425,7 +425,7 @@ impl<D: Digest + BlockSizeUser + FixedOutput, const N: usize> Authentication for
         session: &mut MechanismData,
         input: Option<&[u8]>,
         writer: &mut dyn Write,
-    ) -> Result<(State, Option<usize>), SessionError> {
+    ) -> Result<State, SessionError> {
         use ScramServerState::*;
         match self.state.take() {
             Some(WaitingClientFirst(state)) => {
@@ -436,14 +436,14 @@ impl<D: Digest + BlockSizeUser + FixedOutput, const N: usize> Authentication for
                 let new_state =
                     state.step(&mut rng, session, client_first, writer, &mut written)?;
                 self.state = Some(WaitingClientFinal(new_state));
-                Ok((State::Running, Some(written)))
+                Ok(State::Running)
             }
             Some(WaitingClientFinal(state)) => {
                 let client_final = input.ok_or(SessionError::InputDataRequired)?;
                 let mut written = 0;
                 let new_state = state.step(client_final, session, writer, &mut written)?;
                 self.state = Some(Finished(new_state));
-                Ok((State::Finished, Some(written)))
+                Ok(State::Finished(MessageSent::Yes))
             }
             Some(Finished(_state)) => Err(SessionError::MechanismDone),